9745 - Extract virtualization logic into VirtualizedDocumentList component
